rename service from ibc to cosmos
