+++ web/src/features/CosmosWallet/index.ts
@@ -1,4 +1,4 @@
-import { useIbcChainSelection } from "./hooks/useIbcChainSelection";
-import { sendIbcTransfer } from "./services/ibc";
+import { useCosmosChainSelection } from "./hooks/useCosmosChainSelection";
+import { sendIbcTransfer } from "./services/cosmos";
 
-export { sendIbcTransfer, useIbcChainSelection };
+export { sendIbcTransfer, useCosmosChainSelection };
